Add a dedicated types package
