+++ models/esrgan-legacy/package.json
@@ -2,6 +2,9 @@
   "name": "@upscalerjs/esrgan-legacy",
   "version": "0.3.1",
   "description": "ESRGAN Legacy Models for Upscaler.js",
+  "dependencies": {
+    "upscalerjs-types": "workspace:*"
+  },
   "exports": {
     "./div2k/2x": {
       "require": "./dist/node/div2k/2x.js",
