+++ models/esrgan-legacy/src/div2k/2x.ts
@@ -1,4 +1,4 @@
-import { ModelDefinition } from 'upscaler/types';
+import { ModelDefinition } from 'upscalerjs-types';
 import { NAME, VERSION } from '../constants.generated';
 
 const modelDefinition: ModelDefinition = {
