+++ models/esrgan-slim/tsconfig.json
@@ -19,12 +19,6 @@
     "outDir": "./dist",
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "**/*.test.ts"],
-  
-  "references": [
-    {
-      "path": "../../packages/upscalerjs/tsconfig.json"
-    }
-  ]
+  "exclude": ["node_modules", "**/*.test.ts"]
 }
