+++ models/pixel-upsampler/package.json
@@ -2,6 +2,9 @@
   "name": "@upscalerjs/pixel-upsampler",
   "version": "0.4.5",
   "description": "Pixel Upsampler Model for Upscaler.js",
+  "dependencies": {
+    "upscalerjs-types": "workspace:*"
+  },
   "exports": {
     ".": {
       "require": "./dist/node/2x.js",
