+++ models/pixel-upsampler/tsconfig.json
@@ -17,16 +17,7 @@
     "baseUrl": "./src",
     "rootDir": "./src",
     "outDir": "./dist",
-    "paths": {
-      "upscaler/*": ["../../../packages/upscalerjs/src/*"]
-    }
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", "**/*.test.ts"],
-  
-  "references": [
-    {
-      "path": "../../packages/upscalerjs/tsconfig.json"
-    }
-  ]
+  "exclude": ["node_modules", "**/*.test.ts"]
 }
