+++ packages/upscalerjs/package.json
@@ -82,6 +82,7 @@
     "@tensorflow/tfjs": "^3.12.0"
   },
   "dependencies": {
+    "upscalerjs-types": "workspace:*",
     "tensor-as-base64": "^0.2.0"
   },
   "gitHead": "cccae63a0a7aeccaa78d30e55a8e12ec72f5c19f",
