+++ packages/upscalerjs/src/types.ts
@@ -1,4 +1,5 @@
 import { tf, } from './dependencies.generated';
+import { ModelDefinition, } from 'upscalerjs-types';
 // import { SerializableConstructor, } from '@tensorflow/tfjs-core/dist/serialization';
 
 export type WarmupSizesByPatchSize = {
@@ -34,25 +35,7 @@ export interface UpscaleArgs<P extends Progress<O, PO>, O extends ResultFormat =
   signal?: AbortSignal;
 }
 
-export type ProcessFn<T extends tf.Tensor> = (t: T) => T;
-export interface PackageInformation {
-  name: string;
-  version: string;
-}
-
-type CustomLayer = Parameters<typeof tf.serialization.registerClass>[0];
-export interface ModelDefinition {
-  path: string;
-  scale: 2 | 3 | 4;
-  channels?: 3;
-  packageInformation?: PackageInformation;
-  preprocess?: ProcessFn<tf.Tensor4D>;
-  postprocess?: ProcessFn<tf.Tensor3D>;
-  // customLayers?: SerializableConstructor<tf.layers.Layer>[];
-  customLayers?: CustomLayer[];
-
-  /* eslint-disable @typescript-eslint/no-explicit-any */
-  meta?: Record<string, any>;
-}
 
 export type Layer = tf.layers.Layer;
+
+export { ModelDefinition, PackageInformation, ProcessFn } from 'upscalerjs-types';
