+++ tsconfig.json
@@ -5,15 +5,5 @@
     "target": "ES2020",
     "module": "commonjs",
     "strict": true,
-    "composite": true,
-  },
-  "files": [],
-  "references": [
-    {
-      "path": "./packages/upscalerjs"
-    },
-    {
-      "path": "./models/pixel-upsampler"
-    }
-  ]
+  }
 }
